Přidání obrázků.

+++ chatterbox/views.py
@@ -1,4 +1,5 @@
 from django.contrib.auth.decorators import login_required
+from django.core.files.storage import FileSystemStorage
 from django.forms import ModelForm
 from django.http import HttpResponseRedirect
 from django.shortcuts import render, redirect, HttpResponse
@@ -22,12 +23,20 @@ def home(request):
 
 
 @login_required
-def search(request, s):
-    rooms = Room.objects.filter(name__contains=s)
-    messages = Message.objects.filter(body__contains=s)
+def search(request):
+    if request.method == 'POST':
+        s = request.POST.get('search')
+        s = s.strip()
+        if len(s) > 0:
+            rooms = Room.objects.filter(name__contains=s)
+            messages = Message.objects.filter(body__contains=s)
 
-    context = {'rooms': rooms, 'messages': messages}
-    return render(request, "chatterbox/search.html", context)
+            context = {'rooms': rooms, 'messages': messages, 'search': s}
+            return render(request, "chatterbox/search.html", context)
+        return redirect('home')
+
+    return redirect('home')
+
 
 
 @login_required
@@ -37,18 +46,24 @@ def room(request, pk):
 
     # pokud zadáme novou zprávu, musíme ji zpracovat
     if request.method == 'POST':
+        file_url = ""
+        if request.FILES.get('upload'):                        # pokud jsme poslali soubor
+            upload = request.FILES['upload']               # z requestu si vytáhnu soubor
+            file_storage = FileSystemStorage()             # práce se souborovým systémem
+            file = file_storage.save(upload.name, upload)  # uložíme soubor na disk
+            file_url = file_storage.url(file)              # vytáhnu ze souboru url adresu a uložím
         body = request.POST.get('body').strip()
-        if len(body) > 0:
+        if len(body) > 0 or request.FILES.get('upload'):
             message = Message.objects.create(
                 user=request.user,
                 room=room,
-                body=body
+                body=body,
+                file=file_url                              # vložíme url souboru do databáze
             )
         return HttpResponseRedirect(request.path_info)
 
     context = {'room': room, 'messages': messages}
     return render(request, "chatterbox/room.html", context)
-
 
 @login_required
 def rooms(request):
@@ -75,18 +90,29 @@ def create_room(request):
 
     return render(request, 'chatterbox/create_room.html')
 
+
 @login_required
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
-    room.delete()
+    if room.messages_count() == 0:  # pokud v místnosti není žádná zpráva
+        room.delete()               # tak místnost smažeme
 
+        return redirect('rooms')
+
+    context = {'room': room, 'message_count': room.messages_count()}
+    return render(request, 'chatterbox/delete_room.html', context)
+
+
+
+@login_required
+def delete_room_yes(request, pk):
+    room = Room.objects.get(id=pk)
+    room.delete()
     return redirect('rooms')
 
-    return redirect('home')
 
 
     # formulář
-
 class RoomEditForm(ModelForm):
 
     class Meta:
